requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
PyYAML==6.0.2
scikit-learn==1.5.1
tqdm==4.66.5
torch==2.4.0
mlflow==2.15.1
dvc==3.59.0
apache-airflow==2.10.1
networkx==3.3
joblib==1.4.2